# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Task:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: Template for a RESTful web service in Go with Fiber.
  title: Go REST API Template
  version: 0.0.1
paths:
  /tasks:
    delete:
      description: Delete a Task by its id.
      parameters:
      - description: Id of the Task be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      summary: Delete a Task and, in case there is no Task with the given ID,
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Get all Tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
      summary: Get all Tasks.
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new Task.
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
      summary: Create a new Task.
      tags:
      - Tasks
  /tasks/{id}:
    get:
      description: Get a Task by its id.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "404":
          description: Not Found
      summary: Get a Task by its id.
      tags:
      - Tasks
swagger: "2.0"